+++ src/Domain/ValueObject/StringValue.ts
@@ -1,3 +1,5 @@
-import { SimpleValueObject } from './SimpleValueObject';
+import { SimpleValueObject, SimpleValueObjectConstructor } from './SimpleValueObject';
+
+export type StringValueConstructor<T extends StringValue<any> = StringValue<any>> = SimpleValueObjectConstructor<T, string>;
 
 export abstract class StringValue<T extends StringValue<any> = any> extends SimpleValueObject<T, string> {}