+++ src/Domain/ValueObject/UserId.ts
@@ -2,4 +2,4 @@ import { StringValue } from "./StringValue";
 import { ValueObject } from "./ValueObject";
 
 @ValueObject('Account')
-export class UserId extends StringValue {}
+export class UserId extends StringValue<UserId> {}